Catch AmazonClientException; ensure actor system is shutdown
LayerRegistry creation throws an AmazonClientException when API
credentials cannot be determined. This prevents the service from being
able to interact with Amazon S3.

In addition, the ActorSystem initialization is never shutdown after the
exception, so Spark doesn't know that the job is in a failed state and
can be terminated. This keeps the Spark job alive without a functioning
Spray server instance to process requests.